Fixes and additions to spike

+++ spike/dl_data.py
@@ -1,17 +1,25 @@
 import urllib.request as urllib2
 import zipfile
+import os
 
 url = "https://archive.ics.uci.edu/ml/machine-learning-databases/00344/Activity%20recognition%20exp.zip"
 
 path = "data/data.zip"
 
-print("Downloading 771 mb")
+if not os.path.isfile(path):
 
-urllib2.urlretrieve(url, path)
+    print("Downloading 771 mb")
 
-print("Download complete")
-print("Extracting")
+    urllib2.urlretrieve(url, path)
 
-zip_ref = zipfile.ZipFile(path, 'r')
-zip_ref.extractall("data")
-zip_ref.close()
+    print("Download complete")
+    print("Extracting")
+
+    zip_ref = zipfile.ZipFile(path, 'r')
+    zip_ref.extractall("data")
+    zip_ref.close()
+
+else:
+    
+    print("Data already downloaded")
+
